requirements.txt
streamlit
pandas
mysql-connector-python
faker
plotly
geopy